+++ HISTORY.rst
@@ -30,7 +30,8 @@ End-User Summary
 - Rebuild of variant summary database table happens every Sunday at 2:22am.
 - Added celery queues ``maintenance`` and ``export``.
 - Adding support for connecting two sites via the GAGH Beacon protocol.
-- Adding link-out to "GenCC"
+- Adding link-out to "GenCC".
+- Adding "submit to SPANR" feature.
 
 Full Change List
 ================
@@ -59,7 +60,8 @@ Full Change List
 - Added celery queues ``maintenance`` and ``export``.
 - Adding support for connecting two sites via the GAGH Beacon protocol.
 - Making CADD version behind CADD REST API configurable.
-- Adding link-out to "GenCC"
+- Adding link-out to "GenCC".
+- Adding "submit to SPANR" feature.
 
 -------
 v0.22.1